Se agrego un boton que muestra los mensajes procesados hasta el momento (no se actualiza la tabla en tiempo real)

+++ Corte_03/Taller_hilos_concurrencia/HilosyConcurrencia/src/main/java/nodos/NodoDestino.java
@@ -6,8 +6,8 @@ import java.util.concurrent.ThreadLocalRandom;
 public class NodoDestino{
     
     private double temperatura;
-    private int  humedad;
-    private int luminiscencia;
+    private double  humedad;
+    private double luminiscencia;
     
     public NodoDestino(){}
